gz migration for forward-port (#458)
