+++ src/python_pybind11/src/Interval.cc
@@ -17,7 +17,7 @@
 
 #include "Interval.hh"
 
-namespace ignition
+namespace gz
 {
 namespace math
 {
@@ -31,4 +31,4 @@ void defineMathInterval(py::module &m, const std::string &typestr)
 
 }  // namespace python
 }  // namespace math
-}  // namespace ignition
+}  // namespace gz
